chore(api): move entitie files
