+++ src/shared/api/entities/Base.ts
